desafio_4/requirements.txt
langchain
langchain-ollama
python-dotenv
pandas
openpyxl
python-holidays